Integration tests appsettings fix
Since integratin test project also included the api project it means that both the appsettings from api project and integration test project is copied to output dir.

There is not right now any way to exldue the appsettigns from the api project when publishing the integration test project.

So for now we turned on a flag to not generate error. it will be random which appsettings file makes it through.
However right now all the values in integration test appsettings also exist in api appsettings meaning the values will be correctly replaced.

However this is NOT a long term solution