Change error logging on queue item
Change so we only record direct error message and not complete exception on queue item. Most failures are due to the worker failing the call so the direct response message is usually what is interesting.

We do also on the lines below log the entire exception so we can check the logs for more.