Fix for dubupdate changed provider
After updating dbupdate we were getting errors with certificate.
This was because dbupdate changed provider.

So we updated to trust server certificates (self signed) which was the issue.
This only applies to dev as ConnectionString will be swapped out on deploy for valid string with either valid cert or other solution.